refactor: move code from prepare to classify -> function create_splits and print statements in prepare as function in classify (check_splits)
