+++ src/classify/prepare_data.py
@@ -3,7 +3,6 @@ Prepare data for classifiction
 '''
 import pathlib
 import pandas as pd
-from sklearn.model_selection import train_test_split
 
 import sys 
 sys.path.append(str(pathlib.Path(__file__).parents[2]))
@@ -201,69 +200,6 @@ def filter_metrics(df, percent_NA=0.8, percent_zero=0.8, verbose=True, log_file:
 
     return filtered_metrics
 
-def create_split(df, random_state=129, val_test_size:float=0.15, outcome_col="is_human", feature_cols:list=None, save_path=None, verbose=False):
-    '''
-    Create X, y from df, split into train, test and val 
-
-    Args: 
-        df: dataframe to split
-        random_state: seed for split for reproducibility
-        val_test_size: size of validation and test sets. 0.15 results in 15% val and 15% test. 
-        feature_cols: feature columns in df (predictors)
-                      If None, defaults to all viable features (removing outcome column "is_human" and other irrelevant cols)
-        outcome_col: column for outcome. Defaults to "is_human"
-        verbose: 
-        save_path: directory to save splitted data. If None, does not sav
-
-    Returns: 
-        splits: dict with all splits 
-    '''
-    # take all cols for X if feature_cols is unspecified, otherwise subset df to incl. only feature_cols
-    if feature_cols == None: 
-        cols_to_drop = ["id", "is_human", "dataset", "sample_params", "model", "temperature", "prompt_number", "unique_id"] +  (["annotations"] if "annotations" in df.columns else []) # drop annotation if present (only present for dailydialog)
-        X = df.drop(columns=cols_to_drop)
-    else:
-        X = df[feature_cols]
-
-    # if model col is present, make explicit categorical for xgboost
-    if "model" in X.columns:
-        X["model"] = X["model"].astype("category")
-
-    # subset df to a single outcome col for y 
-    y = df[[outcome_col]]
-
-    splits = {}
-
-    # create train, test, val splits based on val_test_size, save to splits dict. If val_test_size = 0.15, 15% val and 15% test (and stratify by y to keep class balance as much as possible)
-    splits["X_train"], splits["X_test"], splits["y_train"], splits["y_test"] = train_test_split(X, y, test_size=val_test_size*2, random_state=random_state, stratify=y)
-
-    # split val from test, stratify by y again 
-    splits["X_val"], splits["X_test"], splits["y_val"], splits["y_test"] = train_test_split(splits["X_test"], splits["y_test"], test_size=0.5, random_state=random_state, stratify=splits["y_test"])
-  
-    # validate size of splits, print info msg
-    if verbose:
-        # make table of split sizes (absolute numbers and percentages)
-        split_sizes = pd.DataFrame(index=["train", "val", "test"], columns=["absolute", "percentage"])
-
-        # compute absolute numbers, only for X as they should be the same for y
-        total_size = len(df)
-        for split in ["train", "val", "test"]:
-            split_sizes.loc[split, "percentage"] = len(splits[f"X_{split}"]) / total_size * 100
-            split_sizes.loc[split, "absolute"] = len(splits[f"X_{split}"])
-
-        print("\n[INFO:] Split sizes:\n")
-        print(split_sizes)
-        print(f"\nTotal size: {total_size}")
-
-    # save splits to save_path if specified
-    if save_path: 
-        save_path.mkdir(parents=True, exist_ok=True)
-        # get dataset name from df
-        for key, value in splits.items():
-            value.to_csv(save_path / f"{key}.csv")
-
-    return splits
-
 def main(): 
     path = pathlib.Path(__file__)
     datapath = path.parents[2] / "metrics"
@@ -277,17 +213,6 @@ def main():
 
     # filter metrics
     df = filter_metrics(df, percent_NA=0.9, percent_zero=0.9, verbose=True)
-
-    splits = create_split(df, random_state=129, val_test_size=0.15, outcome_col="is_human", verbose=True)
-    
-    # group by dataset and model
-    print("\nPrinting groupby...\n")
-    print(df.groupby(["dataset", "model"]).size())
-
-    print("\nPrinting class distribution... \n")
-    print(f"Y train: {splits['y_train'].value_counts()[0]} AI, {splits['y_train'].value_counts()[1]} human")
-    print(f"Y val: {splits['y_val'].value_counts()[0]} AI, {splits['y_val'].value_counts()[1]} human")
-    print(f"Y test: {splits['y_test'].value_counts()[0]} AI, {splits['y_test'].value_counts()[1]} human")
 
 
 if __name__ == "__main__":
